--- .gitignore
@@ -1,1 +1,0 @@
-AbletonTestModularMadness.xcodeproj
