+++ AbletonTestModularMadness/Model/Module/EchoModule.hpp
@@ -1,22 +1,22 @@
 //
-//  NoopModule.hpp
-//  AbletonTestModularMadness
+//  EchoModule.hpp
+//  src
 //
 //  Created by Sander van Kasteel on 21/11/2018.
 //  Copyright © 2018 Sander van Kasteel. All rights reserved.
 //
 
-#ifndef NoopModule_hpp
-#define NoopModule_hpp
+#ifndef EchoModule_hpp
+#define EchoModule_hpp
 
 #include <stdio.h>
 #include "Module.hpp"
 
-class NoopModule : public Module {
+class EchoModule : public Module {
 public:
-    void process() override;
-
     string process(string in) override;
+
+    void process() override;
 };
 
-#endif /* NoopModule_hpp */
+#endif /* EchoModule_hpp */
