+++ AbletonTestModularMadness/Model/Module/ModuleFactory.hpp
@@ -1,6 +1,6 @@
 //
 //  ModuleFactory.hpp
-//  AbletonTestModularMadness
+//  src
 //
 //  Created by Sander van Kasteel on 21/11/2018.
 //  Copyright © 2018 Sander van Kasteel. All rights reserved.
