+++ AbletonTestModularMadness/Model/Module/ReverseModule.hpp
@@ -1,22 +1,23 @@
 //
-//  EchoModule.hpp
-//  AbletonTestModularMadness
+//  ReverseModule.hpp
+//  src
 //
 //  Created by Sander van Kasteel on 21/11/2018.
 //  Copyright © 2018 Sander van Kasteel. All rights reserved.
 //
 
-#ifndef EchoModule_hpp
-#define EchoModule_hpp
+#ifndef ReverseModule_hpp
+#define ReverseModule_hpp
 
 #include <stdio.h>
 #include "Module.hpp"
 
-class EchoModule : public Module {
+class ReverseModule : public Module {
 public:
-    string process(string in) override;
-
     void process() override;
+
+    string process(string) override;
 };
 
-#endif /* EchoModule_hpp */
+
+#endif /* ReverseModule_hpp */
