+++ AbletonTestModularMadness/Model/Network.cpp
@@ -1,6 +1,6 @@
 //
 //  Network.cpp
-//  AbletonTestModularMadness
+//  src
 //
 //  Created by Sander van Kasteel on 20/11/2018.
 //  Copyright © 2018 Sander van Kasteel. All rights reserved.
