+++ AbletonTestModularMadness/Model/Network.hpp
@@ -1,6 +1,6 @@
 //
 //  Network.hpp
-//  AbletonTestModularMadness
+//  src
 //
 //  Created by Sander van Kasteel on 20/11/2018.
 //  Copyright © 2018 Sander van Kasteel. All rights reserved.
