--- src/Model/Module/ReverseModule.hpp
@@ -1,23 +1,0 @@
-//
-//  ReverseModule.hpp
-//  AbletonTestModularMadness
-//
-//  Created by Sander van Kasteel on 21/11/2018.
-//  Copyright © 2018 Sander van Kasteel. All rights reserved.
-//
-
-#ifndef ReverseModule_hpp
-#define ReverseModule_hpp
-
-#include <stdio.h>
-#include "Module.hpp"
-
-class ReverseModule : public Module {
-public:
-    void process() override;
-
-    string process(string) override;
-};
-
-
-#endif /* ReverseModule_hpp */
